@@ -2,7 +2,7 @@ source 'https://rubygems.org'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.1'
+gem 'rails', '>= 5.0.0.rc2', '< 5.1'
 # Use postgresql as the database for Active Record
 gem 'pg'
 # Use SCSS for stylesheets
@@ -53,7 +53,7 @@ gem 'stripe', :git => 'https://github.com/stripe/stripe-ruby'
 
 gem "paperclip", :git => "git://github.com/thoughtbot/paperclip.git"
 
-gem 'aws-sdk', '< 2.0'
+gem 'aws-sdk', '~> 3'
 
 gem 'activerecord-session_store', github: 'rails/activerecord-session_store'
 
