@@ -1,4 +1,4 @@
-class Admin < ActiveRecord::Base
+class Admin < ApplicationRecord
 	has_secure_password
   validates(:username,    { :uniqueness   => { case_sensitive: false }})
   validates(:password, { :length       => { :minimum => 8, :maximum => 16 },
