@@ -1,4 +1,4 @@
-class Listing < ActiveRecord::Base
+class Listing < ApplicationRecord
 	# has_attached_file :pictures, :styles => { :medium => "300x300>", :thumb => "100x100>" }, :default_url => "/images/:style/missing.png"
   	# validates_attachment_content_type :avatar, :content_type => /\Aimage\/.*\Z/
 	has_many :pictures
